# docs/api_spec.yaml
openapi: 3.1.0
info:
  title: 助盲跑 API
  description: 盲人陪跑志愿者匹配服务
  version: 1.0.0
servers:
- url: http://localhost:8081
  description: 本地开发环境
- url: http://47.114.113.171
  description: 阿里云生产环境（Swagger 已关闭，仅用于请求地址参考）
security:
- Bearer: []
...
paths:
  /api/volunteer/profile:
    get:
      tags:
      - volunteer-controller
      operationId: getProfile
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/VolunteerProfileResponse'
    put:
      tags:
      - volunteer-controller
      operationId: updateProfile
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VolunteerProfileUpdateRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                type: object
...
  /api/blind/profile:
    get:
      tags:
      - 盲人用户
      summary: 获取盲人资料
      operationId: getProfile_1
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/BlindProfileResponse'
    put:
      tags:
      - 盲人用户
      summary: 更新盲人资料
      operationId: updateProfile_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlindProfileUpdateRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                type: object
...
  /api/volunteer/verification:
    post:
      tags:
      - volunteer-controller
      operationId: submitVerification
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: 资质证件文件（图片或PDF）
              required:
              - file
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                type: object
...
  /api/volunteer/location:
    post:
      tags:
      - volunteer-controller
      operationId: updateLocation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VolunteerLocationRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                type: object
      deprecated: true
...
  /api/blind/location:
    post:
      tags:
      - blind-location-controller
      operationId: updateLocation_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlindLocationRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                type: object
...
components:
  schemas:
    VolunteerAvailableTimeSlot:
      type: object
      properties:
        dayOfWeek:
          type: string
          minLength: 1
          pattern: MONDAY|TUESDAY|WEDNESDAY|THURSDAY|FRIDAY|SATURDAY|SUNDAY
        startTime:
          type: string
          example: '14:30:00'
        endTime:
          type: string
          example: '14:30:00'
      required:
      - endTime
      - startTime
    VolunteerProfileUpdateRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 50
          minLength: 2
        availableTimeSlots:
          type: array
          items:
            $ref: '#/components/schemas/VolunteerAvailableTimeSlot'
        acceptsGuideDog:
          type: boolean
        paceRange:
          type: string
          enum:
          - WALK_RUN
          - EASY
          - MODERATE
          - FAST
          - NO_PREFERENCE
        isAvailable:
          type: boolean
          description: 可服务状态（可选，不传则不修改）
...
    BlindProfileUpdateRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 50
          minLength: 2
        runningPace:
          type: string
          maxLength: 20
          minLength: 0
        specialNeeds:
          type: string
          maxLength: 500
          minLength: 0
        visionLevel:
          type: string
          enum:
          - TOTAL_BLIND
          - LOW_VISION
        hasGuideDog:
          type: boolean
        tetherPreference:
          type: string
          enum:
          - TETHER_ROPE
          - ARM_HOLD
          - VERBAL_ONLY
        chatPreference:
          type: string
          enum:
          - PREFER_CHAT
          - PREFER_QUIET
          - NO_PREFERENCE
        defaultPace:
          type: string
          enum:
          - WALK_RUN
          - EASY
          - MODERATE
          - FAST
          - NO_PREFERENCE
...
    VolunteerLocationRequest:
      type: object
      properties:
        latitude:
          type: number
          format: double
          maximum: 90
          minimum: -90
        longitude:
          type: number
          format: double
          maximum: 180
          minimum: -180
        isOnline:
          type: boolean
      required:
      - latitude
      - longitude
...
    BlindLocationRequest:
      type: object
      properties:
        latitude:
          type: number
          format: double
          maximum: 90
          minimum: -90
        longitude:
          type: number
          format: double
          maximum: 180
          minimum: -180
      required:
      - latitude
      - longitude
...
    VolunteerProfileResponse:
      type: object
      properties:
        name:
          type: string
        verificationStatus:
          type: string
        availableTimeSlots:
          type: array
          items:
            $ref: '#/components/schemas/VolunteerAvailableTimeSlot'
        acceptsGuideDog:
          type: boolean
        paceRange:
          type: string
          enum:
          - WALK_RUN
          - EASY
          - MODERATE
          - FAST
          - NO_PREFERENCE
        isAvailable:
          type: boolean
          description: 可服务状态，false 时不参与派单也不能接单
...
    BlindProfileResponse:
      type: object
      properties:
        name:
          type: string
        runningPace:
          type: string
        specialNeeds:
          type: string
        verifyStatus:
          type: string
          enum:
          - NOT_VERIFIED
          - VERIFIED
          - FAILED
        visionLevel:
          type: string
          enum:
          - TOTAL_BLIND
          - LOW_VISION
        hasGuideDog:
          type: boolean
        tetherPreference:
          type: string
          enum:
          - TETHER_ROPE
          - ARM_HOLD
          - VERBAL_ONLY
        chatPreference:
          type: string
          enum:
          - PREFER_CHAT
          - PREFER_QUIET
          - NO_PREFERENCE
        defaultPace:
          type: string
          enum:
          - WALK_RUN
          - EASY
          - MODERATE
          - FAST
          - NO_PREFERENCE
  securitySchemes:
    Bearer:
      type: http
      name: Authorization
      scheme: bearer
      bearerFormat: JWT